Add Kuma Report template generator
Add a Python script to generate the template for the monthly Kuma
report. This uses:

* requests (popular HTTP requests library)
* requests-cache (persistent cache for requests)
* Credentials for a GitHub OAuth app, created by the user

It parses the PR log for several MDN projects, to find the merged PRs
for the previous month, and which are from first-time contributors. It
uses this data to create a template for the monthly Kuma report, and
prints it to stdout.